Bulk Load CDK: ObjectStorageStreamLoader
