+++ airbyte-cdk/bulk/toolkits/load-s3/src/main/kotlin/io/airbyte/cdk/load/file/s3/S3Client.kt
@@ -15,8 +15,6 @@ import aws.smithy.kotlin.runtime.content.ByteStream
 import aws.smithy.kotlin.runtime.content.toInputStream
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings
 import io.airbyte.cdk.load.command.aws.AWSAccessKeyConfigurationProvider
-import io.airbyte.cdk.load.command.object_storage.ObjectStorageCompressionConfiguration
-import io.airbyte.cdk.load.command.object_storage.ObjectStorageCompressionConfigurationProvider
 import io.airbyte.cdk.load.command.object_storage.ObjectStorageUploadConfiguration
 import io.airbyte.cdk.load.command.object_storage.ObjectStorageUploadConfigurationProvider
 import io.airbyte.cdk.load.command.s3.S3BucketConfiguration
@@ -25,7 +23,6 @@ import io.airbyte.cdk.load.file.NoopProcessor
 import io.airbyte.cdk.load.file.StreamProcessor
 import io.airbyte.cdk.load.file.object_storage.ObjectStorageClient
 import io.airbyte.cdk.load.file.object_storage.RemoteObject
-import io.github.oshai.kotlinlogging.KotlinLogging
 import io.micronaut.context.annotation.Factory
 import io.micronaut.context.annotation.Secondary
 import jakarta.inject.Singleton
@@ -45,9 +42,7 @@ class S3Client(
     private val client: aws.sdk.kotlin.services.s3.S3Client,
     val bucketConfig: S3BucketConfiguration,
     private val uploadConfig: ObjectStorageUploadConfiguration?,
-    private val compressionConfig: ObjectStorageCompressionConfiguration<*>? = null,
 ) : ObjectStorageClient<S3Object> {
-    private val log = KotlinLogging.logger {}
 
     override suspend fun list(prefix: String) = flow {
         var request = ListObjectsRequest {
@@ -125,23 +120,12 @@ class S3Client(
     }
 
     override suspend fun delete(key: String) {
-        val request = DeleteObjectRequest {
-            bucket = bucketConfig.s3BucketName
-            this.key = key
-        }
-        client.deleteObject(request)
-    }
-
-    override suspend fun streamingUpload(
-        key: String,
-        block: suspend (OutputStream) -> Unit
-    ): S3Object {
-        return streamingUpload(key, compressionConfig?.compressor ?: NoopProcessor, block)
+        delete(S3Object(key, bucketConfig))
     }
 
     override suspend fun <U : OutputStream> streamingUpload(
         key: String,
-        streamProcessor: StreamProcessor<U>,
+        streamProcessor: StreamProcessor<U>?,
         block: suspend (OutputStream) -> Unit
     ): S3Object {
         val request = CreateMultipartUploadRequest {
@@ -154,7 +138,7 @@ class S3Client(
                 client,
                 response,
                 ByteArrayOutputStream(),
-                streamProcessor,
+                streamProcessor ?: NoopProcessor,
                 uploadConfig
             )
         upload.runUsing(block)
@@ -167,7 +151,6 @@ class S3ClientFactory(
     private val keyConfig: AWSAccessKeyConfigurationProvider,
     private val bucketConfig: S3BucketConfigurationProvider,
     private val uploadConifg: ObjectStorageUploadConfigurationProvider? = null,
-    private val compressionConfig: ObjectStorageCompressionConfigurationProvider<*>? = null,
 ) {
     companion object {
         fun <T> make(config: T) where
@@ -195,7 +178,6 @@ class S3ClientFactory(
             client,
             bucketConfig.s3BucketConfiguration,
             uploadConifg?.objectStorageUploadConfiguration,
-            compressionConfig?.objectStorageCompressionConfiguration,
         )
     }
 }
