+++ airbyte-integrations/connectors/destination-s3-v2/src/test-integration/resources/expected-spec-cloud.json
@@ -297,7 +297,7 @@
       "s3_staging_prefix" : {
         "type" : "string",
         "default" : "{s3_bucket_path}/__airbyte_tmp",
-        "description" : "Path to use when staging data in the bucket directory. Documentation TBD.",
+        "description" : "Path to use when staging data in the bucket directory. Airbyte will stage data here during sync and/or write small manifest/recovery files.",
         "title" : "S3 Staging Prefix",
         "examples" : [ "__staging/data_sync/test" ]
       }
